Added sprint component
